refactor bandcamp search functionality and delete unused vars
